% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_lgb_deviance_func}
\alias{cidx_lgb_deviance_func}
\title{Cindex deviance for Lgb}
\usage{
cidx_lgb_deviance_func(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
deviance value
}
\description{
This function calculates cindex deviance for lgb model.
}
