% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_lgb_func}
\alias{cidx_lgb_func}
\title{C-index for Lgb}
\usage{
cidx_lgb_func(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
cindex value
}
\description{
This function evaluates C-index for lgb model.
}
