% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_lgb_obj}
\alias{cidx_lgb_obj}
\title{Cindex objective func for Lgb}
\usage{
cidx_lgb_obj(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
gradient and hessian
}
\description{
This function calculates cindex objcective function for lgb model.
}
