% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_xgb_func}
\alias{cidx_xgb_func}
\title{C-index for xgb}
\usage{
cidx_xgb_func(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
cindex value
}
\description{
This function evaluates C-index for xgb model.
}
