% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_xgb_obj}
\alias{cidx_xgb_obj}
\title{Cindex objective func for xgb}
\usage{
cidx_xgb_obj(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
gradient and hessian
}
\description{
This function calculates cindex objcective function for xgb model.
}
