% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{Cox_lgb_deviance_func}
\alias{Cox_lgb_deviance_func}
\title{define objective function and evaluation function}
\usage{
Cox_lgb_deviance_func(preds, dtrain)
}
\description{
define objective function and evaluation function
}
\details{
@export Cox_lgb_deviance_func
}
