% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{Cox_lgb_obj}
\alias{Cox_lgb_obj}
\title{Cox objective func for Lgb}
\usage{
Cox_lgb_obj(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
gradient and hessian
}
\description{
This function calculates cox objcective function for lgb model.
}
