% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{Cox_xgb_deviance_func}
\alias{Cox_xgb_deviance_func}
\title{Cox deviance for xgb}
\usage{
Cox_xgb_deviance_func(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
deviance value
}
\description{
This function calculates cox deviance for xgb model.
}
