% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_trans.R
\name{leaf_tf}
\alias{leaf_tf}
\title{Transform feature space to high dimension leaf space for xgb or lgb model}
\usage{
leaf_tf(model, datax)
}
\arguments{
\item{model}{model object}

\item{datax}{X data set}

\item{datay}{Y data set}
}
\value{
a one hot-fashion feature space
}
\description{
This function transforms the feature space to a high dimension leaf space
}
