% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfit.R
\name{lgb.sur}
\alias{lgb.sur}
\title{Fitting lgb model}
\usage{
lgb.sur(
  datax,
  datay,
  method = c("defaut", "pl", "C"),
  nfolds = 5,
  nround = NULL,
  lambda = NULL,
  alpha = NULL,
  eta = NULL,
  early_stopping_rounds = NULL
)
}
\arguments{
\item{datax}{X data set}

\item{datay}{Y data set}

\item{method}{fitting method}

\item{nfolds}{number of folds}

\item{nround}{number of rounds}

\item{lambda}{l1 parameter}

\item{alpha}{l2 parameter}

\item{eta}{learning rate}

\item{early_stopping_rounds}{force a stopping round}
}
\value{
a fitted xgb model
}
\description{
This function allows you to fit survival data with a lgb model.
}
