% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_surv_tree.R
\name{sim_surv_tree}
\alias{sim_surv_tree}
\title{Simple survival tree}
\usage{
sim_surv_tree(model, x_data, y_data, top_nc = NULL, maxdp = NULL, cp = NULL)
}
\arguments{
\item{model}{lightgbm model object}

\item{x_data}{X data set}

\item{y_data}{Y data set}

\item{top_nc}{numbers of top features selected}

\item{maxdp}{max depth for rpart tree}

\item{cp}{complexity param for rpart tree}
}
\value{
a list contains 1.rpart tree;2.ctree
}
\description{
This function allows you to find risk level automatically with xgbor lgb object.
}
