% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_risk.R
\name{surv_risk_aut}
\alias{surv_risk_aut}
\title{Automatically define survival risk with xgb or lgb model}
\usage{
surv_risk_aut(model, train_data, test_data)
}
\arguments{
\item{model}{fitted model (xgb or lgb)}

\item{train_data}{train data (only covariates)}

\item{test_data}{test data (only covaraites)}
}
\value{
a factor vector with three different risk levels
}
\description{
Find the risk level for patient under model
}
