% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_risk.R
\name{surv_risk_aut_rf}
\alias{surv_risk_aut_rf}
\title{Automatically define survival risk with rf model}
\usage{
surv_risk_aut_rf(model, train_data, test_data)
}
\arguments{
\item{model}{fitted rf model}

\item{train_data}{train data (only covariates)}

\item{test_data}{test data (only covaraites)}
}
\value{
a factor vector with three different risk levels
}
\description{
Find the risk level for patient under a rf model
}
