% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_risk.R
\name{surv_risk_m}
\alias{surv_risk_m}
\title{Manually define survival risk with xgb or lgb model}
\usage{
surv_risk_m(model, train_data, test_data, upper, lower)
}
\arguments{
\item{model}{fitted model (xgb or lgb)}

\item{train_data}{train data (only covariates)}

\item{test_data}{test data (only covaraites)}

\item{upper}{upper percentil cutting point}

\item{lower}{lower percentil cutting point}
}
\value{
a factor vector with three different risk levels
}
\description{
Find the risk level for patient under model
}
