% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xsurv_predict.R
\name{Xsurv_predict_sv}
\alias{Xsurv_predict_sv}
\title{This function allows you to predict a patient's survival curve}
\usage{
Xsurv_predict_sv(model, x_train, y_train, x_new, nc = 25)
}
\arguments{
\item{model}{xgboost or lightgbm model object}

\item{x_train}{X train data set}

\item{y_train}{Y trian data set}

\item{nc}{number of cutting points defaut is 25, and 20-30 is recommended}

\item{x_test}{X test data set}

\item{nq}{quantile of train data time in each cut group for prediction time}
}
\value{
predicted survival probabilty
}
\description{
This function allows you to predict a patient's survival curve
}
