% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xsurv_sim_data.R
\name{Xsurv_sim_data}
\alias{Xsurv_sim_data}
\title{This function allows you to find risk level automatically with lgbcv object.}
\usage{
Xsurv_sim_data(size, dim, lambda = NULL, vu = 1, beta = NULL, c_rate)
}
\arguments{
\item{size}{sample size of the simulation data}

\item{dim}{dimension of parameters}

\item{lambda}{scale parameter}

\item{vu}{shape parameter,defaut is 1 for exponential distribution}

\item{beta}{coefficients}

\item{c_rate}{approximate censor rate}
}
\value{
a simulated dataframe
}
\description{
This function allows you to find risk level automatically with lgbcv object.
}
\examples{
#
sim_dat<-Xsurv_sim_data(size=500,dim=20,lambda=2,vu=1,
                                    c_rate=0.3)


sim_x<-sim_dat[,1:20]
sim_y<-sim_dat[,c(21,22)]

fit<-Xsurv.cv(sim_x,sim_y,top_n=5)
}
