% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_lgb_func}
\alias{cidx_lgb_func}
\title{define objective function and evaluation function}
\usage{
cidx_lgb_func(preds, dtrain)
}
\description{
define objective function and evaluation function
}
\details{
@export cidx_lgb_func
}
