% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{cidx_xgb_func}
\alias{cidx_xgb_func}
\title{define objective function and evaluation function}
\usage{
cidx_xgb_func(preds, dtrain)
}
\description{
define objective function and evaluation function
}
\details{
@export cidx_xgb_func
}
