% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{Cox_lgb_deviance_func}
\alias{Cox_lgb_deviance_func}
\title{Cox deviance for Lgb}
\usage{
Cox_lgb_deviance_func(preds, dtrain)
}
\arguments{
\item{preds}{predictor}

\item{dtrain}{train data}
}
\value{
deviance value
}
\description{
This function calculates cox deviance for lgb model.
}
