% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_func.R
\name{Cox_lgb_obj}
\alias{Cox_lgb_obj}
\title{define objective function and evaluation function}
\usage{
Cox_lgb_obj(preds, dtrain)
}
\description{
define objective function and evaluation function
}
\details{
@export Cox_lgb_obj
}
